@Scheduled вынесен в ScheduledTasks
