+++ build.gradle
@@ -7,19 +7,8 @@ plugins {
 group 'ru.mts'
 version '1.0-SNAPSHOT'
 
-//application {
-//    mainModule = 'src.main' // name defined in module-info.java
-//    mainClass = 'src.main.java.ru.mts.hw_3.Main.java'
-//}
-//tasks.withType(Jar) {
-//    duplicatesStrategy = DuplicatesStrategy.EXCLUDE
-//
-//    manifest {
-//        attributes["Main-Class"] = "ru.mts.hw_3.Main"
-//    }
-//}
 jar {
-    manifest.attributes("Main-Class": "ru.mts.hw_3.Main");
+    manifest.attributes('Main-Class': 'ru.mts.hw_3.Main');
 }
 repositories {
     mavenCentral()
